Workaround for primary key type annotation with Py 3.11
In Python 3.10+ type annoations need checking in a more complex way. This is used to determine if a column in a model is a primary key.

This is a minimal, likely incorrect/imperfect fix intended as a workaround until this can be fixed upstream.

See https://github.com/hyperflask/sqlorm/issues/1.